+++ ProjectX/Models/Account.swift
@@ -27,9 +27,9 @@ struct Account: Identifiable, Equatable {
     var maximumLoss: Double
     var lockoutReason: String?
     var lockoutExpiration: String?
-    var personalDailyProfitTarget: Double?
+    var personalDailyProfitTarget: Int?
     var personalDailyProfitTargetAction: Int
-    var personalDailyLossLimit: Double?
+    var personalDailyLossLimit: Int?
     var personalDailyLossLimitAction: Int
     var personalDailyLossLimitTrailing: Bool
     var pdllTrailingType: Int
